refactor(scheduler): drop m_ prefix from classes and loading modules
