+++ packages/devextreme/js/__internal/scheduler/workspaces/m_work_space.ts
@@ -57,9 +57,6 @@ import type { GroupOrientation, ViewType } from '@ts/scheduler/types';
 import Scrollable, { type ScrollableProperties } from '@ts/ui/scroll_view/scrollable';
 
 import type NotifyScheduler from '../base/widget_notify_scheduler';
-import { APPOINTMENT_SETTINGS_KEY } from '../constants';
-import { Cache } from '../global_cache';
-import AppointmentDragBehavior from '../m_appointment_drag_behavior';
 import {
   APPOINTMENT_DRAG_SOURCE_CLASS,
   DATE_TABLE_CLASS,
@@ -69,7 +66,10 @@ import {
   TIME_PANEL_CLASS,
   VERTICAL_GROUP_COUNT_CLASSES,
   VIRTUAL_CELL_CLASS,
-} from '../m_classes';
+} from '../classes';
+import { APPOINTMENT_SETTINGS_KEY } from '../constants';
+import { Cache } from '../global_cache';
+import AppointmentDragBehavior from '../m_appointment_drag_behavior';
 import { CompactAppointmentsHelper } from '../m_compact_appointments_helper';
 import type { SubscribeKey, SubscribeMethods } from '../m_subscribes';
 import tableCreatorModule, { type GroupRows } from '../m_table_creator';
