+++ packages/devextreme/js/__internal/scheduler/workspaces/m_work_space_grouped_strategy_horizontal.ts
@@ -1,7 +1,7 @@
 import { getBoundingRect } from '@js/core/utils/position';
 import { WORK_SPACE_BORDER_PX } from '@ts/scheduler/workspaces/const';
 
-import { FIRST_GROUP_CELL_CLASS, LAST_GROUP_CELL_CLASS } from '../m_classes';
+import { FIRST_GROUP_CELL_CLASS, LAST_GROUP_CELL_CLASS } from '../classes';
 
 class HorizontalGroupedStrategy {
   // TODO: make private once external usages in current_time_shader.ts, current_time_shader_horizontal.ts are removed
