+++ packages/devextreme/js/__internal/scheduler/workspaces/m_work_space_grouped_strategy_vertical.ts
@@ -2,8 +2,8 @@ import { getBoundingRect } from '@js/core/utils/position';
 import { calculateDayDuration, getVerticalGroupCountClass } from '@ts/scheduler/r1/utils/index';
 import { WORK_SPACE_BORDER_PX } from '@ts/scheduler/workspaces/const';
 
+import { FIRST_GROUP_CELL_CLASS, LAST_GROUP_CELL_CLASS } from '../classes';
 import { Cache } from '../global_cache';
-import { FIRST_GROUP_CELL_CLASS, LAST_GROUP_CELL_CLASS } from '../m_classes';
 
 class VerticalGroupedStrategy {
   cache = new Cache();
