+++ packages/devextreme/js/__internal/scheduler/workspaces/work_space_indicator.ts
@@ -9,7 +9,7 @@ import { dateUtilsTs } from '@ts/core/utils/date';
 import type { OptionChanged } from '@ts/core/widget/types';
 import { getToday } from '@ts/scheduler/r1/utils/index';
 
-import { HEADER_CURRENT_TIME_CELL_CLASS } from '../m_classes';
+import { HEADER_CURRENT_TIME_CELL_CLASS } from '../classes';
 import timezoneUtils from '../m_utils_time_zone';
 import SchedulerWorkSpace, {
   type WorkspaceOptionsInternal,
