+++ packages/devextreme/testing/tests/DevExpress.ui.widgets.scheduler/loading.tests.js
@@ -1,5 +1,5 @@
 const $ = require('jquery');
-const loading = require('__internal/scheduler/m_loading');
+const loading = require('__internal/scheduler/loading');
 const viewPort = require('core/utils/view_port').value;
 const fx = require('common/core/animation/fx');
 const LoadPanel = require('ui/load_panel');
@@ -28,6 +28,7 @@ QUnit.test('hide loadPanel', async function(assert) {
     });
 
     loading.hide();
+    await Promise.resolve();
     assert.equal($('.dx-loadpanel').length, 0);
 });
